no_update field
If no_update = True for a certain webpage, the content is not updated on change aka the next time the page is checked, it is still compared to a fixed initial content. This is similar to our first website bot that send messages as long as there was a flat online - a pretty good feature, not a bug.
By default, it is disabled (no_update = False). I only set it for the living science page. I set it manually (downloading the webpages file from the server, reading it in the program, adding the no_update field for every existing webpage, setting it for every one except living science to false). Could you @niklasbogensperger implement an admin command to set this flag for a certain webpage? (E.g. /toggle_no_update livingscience changes the boolean value to it's inverse and prints afterwards the /getpageinfo for this very page as a confirmation. Note: I inclueded in the __string__ constructor of the webpage class already the no update info, so there is no need to change anything, just call the method getpage info or str(myWebpage) ). Speaking of this feature, woud it be possible to offer a "close dialoge" or "exit" button for the info that is printed for a certain webpage after calling /getpageinfo myWebpage as there is for the subscibe menu. Is this (with a small amount of work) possible?