enhanced get routine
added header info. Any https should work now. verify = False is just for certificate checking and has nothing to do with https support (the latter is handled by requests in the background). It is intended to have verify=False as otherwise expired certificates would caus an error.
Added the requeststest.py to test any changes to the parameters / headers. Please keep this up to date with the main driver